tests/wpt/web-platform-tests/css/css-toggle/parsing/toggle-visibility-computed.tentative.html

<!DOCTYPE HTML>
<meta charset="UTF-8">
<title>CSS Toggles: getComputedStyle() for toggle-visibility property</title>
<link rel="author" title="L. David Baron" href="https://dbaron.org/">
<link rel="author" title="Google" href="http://www.google.com/">
<link rel="help" href="https://tabatkins.github.io/css-toggle/#toggle-visibility-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>

<body>
<div id="target"></div>
<script>

test_computed_value('toggle-visibility', 'normal');
test_computed_value('toggle-visibility', 'toggle t');
test_computed_value('toggle-visibility', 'toggle toggle');

</script>
